backend/requirements.txt
fastapi
uvicorn
yfinance
numpy
scipy
pandas
openai
python-dotenv
httpx
pydantic